ADD : transaction controller/router/model

+++ controller/person.controller.js
@@ -1,12 +1,9 @@
 const Person = require("../model/person.model");
+const User = require("../model/user.model");
+const bcrypt = require("bcrypt");
 
 exports.add = async (req, res) => {
     try {
-        // const email = req.body.email;
-        // console.log("::email::", email);
-        // const checkEmail = await Person.findOne({ email: email })
-        // console.log("::checkEmail::", checkEmail);
-
         /* For Generating Unique Code */
         let chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890"
         let uniqueId = "";
@@ -16,7 +13,7 @@ exports.add = async (req, res) => {
         }
 
         const ID = req.user._id
-        const { firstName, lastName, email, reference, } = req.body
+        const { firstName, lastName, email, reference } = req.body
         if (firstName.trim().length == 0 || lastName.trim().length == 0 || email.trim().length == 0 || reference.trim().length == 0) {
             res.status(401).json({
                 message: "PLEASE ENTER ALL FILED",
@@ -41,7 +38,6 @@ exports.add = async (req, res) => {
                 })
 
                 const saveData = await insertData.save();
-                console.log("::saveData::", saveData);
 
                 res.status(201).json({
                     message: "PERSON DATA INSERT SUCCESSFULLY",
@@ -51,7 +47,7 @@ exports.add = async (req, res) => {
             }
         }
     } catch (error) {
-        console.log("person-add-ERROR__:: ", error);
+        console.log("personAdd--ERROR:: ", error);
         res.status(500).json({
             message: "SOMETHING WENT WRONG",
             status: 500
@@ -62,12 +58,13 @@ exports.add = async (req, res) => {
 
 exports.edit = async (req, res) => {
     try {
-        const ID = req.user._id
-        console.log("::ID::", ID);
+        const id = req.user._id;
+        const personId = req.params.id;
 
-        const updateData = await Person.findByIdAndUpdate(
+        const updateData = await Person.updateOne(
             {
-                _id: ID
+                userId: id,
+                personId: personId
             },
             {
                 $set: {
@@ -77,19 +74,16 @@ exports.edit = async (req, res) => {
                     email: req.body.email,
                     reference: req.body.reference
                 }
-            },
-            {
-                new: true
             })
 
-        console.log("::updateData::", updateData);
-
         res.status(200).json({
-            message: "PERSON DETALIS UPDATE SUCCESSFULLY",
-            status: 200
+            message: `${personId}'s DETAILS UPDATE SUCCESSFULLY`,
+            status: 200,
+            data: id
         })
+
     } catch (error) {
-        console.log("person-edit-ERROR__:: ", error);
+        console.log("personEdit--ERROR:: ", error);
         res.status(500).json({
             message: "SOMETHING WENT WRONG",
             status: 500
@@ -100,22 +94,32 @@ exports.edit = async (req, res) => {
 
 exports.deleteData = async (req, res) => {
     try {
-        const ID = req.user._id;
-        console.log("::ID::", ID);
+        const personId = req.params.personId
+        const password = req.body.password
+        const data = req.user._id;
 
-        const deletePerson = await Person.deleteOne(
-            {
-                personId: req.params.personId
-            }
-        );
+        const getUser = await User.findOne({ _id : data})
+        const comparePass = await bcrypt.compare(password, getUser.password);
 
-        res.status(200).json({
-            message: "PERSON DELETE SUCCESSFULLY",
-            status: 200
-        })
+        if (comparePass == true) {
+                const deletePerson = await Person.deleteOne(
+                    {
+                        personId: personId
+                    }
+                );
+                res.status(200).json({
+                    message: "PERSON DELETE SUCCESSFULLY",
+                    status: 200
+                })
+        } else {
+            res.status(401).json({
+                message: "PASSWORD INCORRECT",
+                status: 401
+            })
+        }
 
     } catch (error) {
-        console.log("person-delete-ERROR__:: ", error);
+        console.log("personDelete--ERROR:: ", error);
         res.status(500).json({
             message: "SOMETHING WENT WRONG",
             status: 500
@@ -125,26 +129,26 @@ exports.deleteData = async (req, res) => {
 // ---------- End Delete Person API For Person ---------- //
 
 exports.listByUserId = async (req, res) => {
-        try {
-            const ID = req.user.id
-            const page = req.body.page;
-            const limit = req.body.limit;
+    try {
+        const ID = req.user.id
+        const page = req.body.page;
+        const limit = req.body.limit;
 
-            const getBook = await Person.find({ userId: ID }).limit(limit * 1).skip((page - 1) * limit);
+        const getBook = await Person.find({ userId: ID }).limit(limit * 1).skip((page - 1) * limit);
 
-            res.status(200).json({
-                message: "GET ALL BOOK BY USER",
-                status: 200,
-                page: page,
-                size: limit,
-                data: getBook
-            });
-        } catch (error) {
-            console.log("person-listByUser-ERROR__:: ", error);
-            res.status(500).json({
-                message: "SOMETHING WENT WRONG",
-                status: 500
-            })
-        }
-    };
+        res.status(200).json({
+            message: "GET ALL BOOK BY USER",
+            status: 200,
+            page: page,
+            size: limit,
+            data: getBook
+        });
+    } catch (error) {
+        console.log("personListByUser--ERROR:: ", error);
+        res.status(500).json({
+            message: "SOMETHING WENT WRONG",
+            status: 500
+        })
+    }
+};
 // ---------- End list By User ID API For Person ---------- //
